# docs/swagger.yaml
openapi: 3.0.1
info:
  title: Barbearia - API de Agendamento
  version: 1.0.0
  description: API REST para agendamento de barbearia. Inclui autenticação JWT, CRUD de agendamentos, listagem de barbeiros e serviços.
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
  /api/auth/cadastro:
    post:
      summary: Cadastro de novo usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nome,email,senha]
              properties:
                nome: { type: string }
                email: { type: string }
                senha: { type: string }
      responses:
        '201':
          description: Usuário criado com sucesso
        '400':
          description: Dados inválidos
        '409':
          description: Email já cadastrado
  /api/auth/login:
    post:
      summary: Autenticação e retorno do token JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email,senha]
              properties:
                email: { type: string }
                senha: { type: string }
      responses:
        '200':
          description: Login bem-sucedido — retorna token
        '400':
          description: Dados inválidos
        '401':
          description: Credenciais inválidas
...